relocate files + minor changes
