+++ src/renderer/src/components/editor/EntryNameDialog.tsx
