+++ src/renderer/src/components/editor/editor.tsx
@@ -1,0 +1,1 @@
+export { default } from "./WorkspaceEditor";
