+++ src/renderer/src/components/editor/utils/types.ts
