--- src/renderer/src/components/install/editor.tsx
@@ -1,1 +1,0 @@
-export { default } from "./workspace-editor/WorkspaceEditor";
